Rename reply message to selected message preview
